refactor: move `src/test/schema.ts` file to `src/test/schema/raw.ts`
